requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
pymodbus==2.5.2
requests==2.31.0
structlog==21.1.0
Werkzeug==2.2.3
gpiozero~=1.6.2
setuptools~=65.5.1
schema~=0.7.5
beautifulsoup4~=4.10.0
numpy~=1.22.3
waitress~=2.1.2
smbus2~=0.4.2
PyJWT~=2.4.0
bcrypt~=4.1.3
schedule
Flask
pytest
pytest-cov